requirements.txt
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-magic==0.4.27
flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
werkzeug==2.3.7
python-magic-bin==0.4.14; sys_platform == 'win32'
pydantic==2.5.2
python-multipart==0.0.6 